Hoist get_local_host into imbue.mngr.api.providers (dedupe 4 copies)
The rsync deploy-file fix added a fourth inline copy of the
"resolve the local host as an OnlineHostInterface" helper. It already
existed verbatim in cli.headless_runner, mngr_claude, and
mngr_uncapped_claude. Consolidate them into a single canonical
get_local_host() in imbue.mngr.api.providers (alongside
get_provider_instance, which all callers already depend on), and route
every call site through it:

- cli.headless_runner: removed its get_local_host; cli.ask and
  mngr_mapreduce now import from api.providers.
- mngr_claude / mngr_uncapped_claude: dropped their private
  _get_local_host and now use the shared helper.
- mngr_recursive._upload_deploy_files: uses get_local_host instead of
  the inline get_provider_instance(...).get_host(...) + isinstance check.

No behavior change. The cli.headless_runner location was also a layering
smell -- plugins shouldn't reach into the cli package for this.
