+++ libs/mngr_mapreduce/changelog/mngr-ssh-reset.md
@@ -1,0 +1,1 @@
+Updated the `get_local_host` import to its new canonical home in `imbue.mngr.api.providers` (it previously lived in `imbue.mngr.cli.headless_runner`). No behavior change.
