+++ libs/mngr_uncapped_claude/changelog/mngr-ssh-reset.md
@@ -1,0 +1,1 @@
+Replaced the module-local `_get_local_host` helper with the shared `get_local_host` from `imbue.mngr.api.providers` (deduplication; no behavior change).
